chore: Move generic warning
