+++ apps/web/src/views/Swap/components/SwapWarningModal/GenericTokenWarning.tsx
